Update display of types in help text
The types and defaults are now displayed on a separate line.
This prevents the last line from getting unreasonably long
(since we're using the raw text formatter) and makes it easier
to spot type information when reading the help text.

Defaults are added for all action types except _AppendAction.
This restores the default display for boolean flags.

Examples in the documentation have been updated.