Replace underscores by dashes in subcommand names.
Note that if one of the subcommand functions has a name starting with an
underscore, an invalid parser will be generated, as the corresponding
subcommand will start with a dash and thus not be triggerable from the
command line.  However, I consider that it should be up to argparse to
raise an exception in that case -- if it doesn't do that validation, we
shouldn't bother either.