Fix subcommand summary help.
Previously subcommand help would not display the first line of the
docstring.  The test was incorrect because it checked whether "foo" was
in
```
usage: python -m unittest [-h] {foo,bar} ...

positional arguments:
  {foo,bar}
    foo
    bar

optional arguments:
  -h, --help  show this help message and exit

```
but of course "foo" is present, but as the command name itself.  Change
by checking for another substring.