Suppress traceback for exceptions listed with :raises:.
See new example and documentation.

I chose to ignore the message regarding the exception in the docstring
(if any) as it would typically read less nicely (":raises FooError: if
bar" => "if bar").

I don't *actually* test that the traceback is suppressed because that
traceback is generated when the exception causes terminal unwinding of
the interpreter, so that's a bit of a pain to check unless running a
subprocess.

I bumped the napoleon dependency to 0.7 as that's needed for proper
handling of Raises block in numpydoc format.